+++ src/container_tri.hh
@@ -60,7 +60,7 @@ class container_triclinic_base : public unitcell,  public voro_base_3d {
         int oxyz;
         /** This array holds the numerical IDs of each particle in each
          * computational box. */
-        int **id;
+        uint64_t **id;
         /** A two-dimensional array holding particle positions. For the derived
          * container_poly class, this also holds particle radii. */
         double **p;
